pwrite_s: the checking return value of pwrite drives in a three state condition
The good result is that written bytes are equal with the count input parameter.
But if the result is -1 then this is a clear situation, but when we wrote
some bytes, but not all of them we should set the errno to ENOSPC.

We use pwrite in some places in syslog-ng, therefore a wrapper function is a clean
solution to check this three state condition, to make it easier/cleaner using the pwrite.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>