persist-state: If the increasing of size of persist file is failed we have to stop.
In case of creating persist file fails we should shutdown normally, not with SIGQUIT

In other cases (allocate fails) we should kill with SIGQUIT

If we would not stop, then a lot of undefined behaviors should come,
so the cleanest way is to stop the process.

why is pwrite_s instead of posix_fallocate?

Unfortunately, the posix_fallocate is not available on every supported platforms
There are some platforms, where the posix_fallocate is available,
not works correctly e.g.: on solaris it returns with invalid argument,
which can occur in case of some kind of file-systems.

So the solution is to fill the whole increased space of the persist file with 0.
It is slower, but this is a very rare event.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>