hashtree: added
HashTree is a tree data structure where the children for a node
are stored in a HashTable.
Every node has a key (which is a string) and a value (gpointer).

The tree is created for storing/accessing hierarhical data that can be
addressed via a path. A path is a serie of keys which are separated by
a delimiter string (eg.: "destination::java::elasticsearch").

Lookup cost is proportional to the length of the path (number of keys),
because the amortized cost for the hashtable lookup is 1.

Insertion cost is the same: find longest path(number of keys) then insert the
remain keys (amortized cost of inserting one element to a hashtable is also 1).

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>