cfg: don't call host_resolve_options_init() from cfg_init
This was actually a noop, as cfg->host_resolve_options was initialized
based on cfg->host_resolve_options. So remove that instead.

I was thinking about introducing a "init host_resolve_options
without inheritance", but that would only be an empty function, so
I decided against that. I could probably be convinced both ways.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>