msg-format: make sd_param_value maximum length adjustable via API
Although this option is not (yet) introduced in the configuration file,
the unit tests also try to cover the 'too-long-sdata' case, which is
inpractical in case the new 64k default is to be overflown.

It might make sense to make this value actually controllable by
the configuration file, but for now I think it is enough to simply bump
the default limit to 64k.

This patch also fixes the test_msgparse unit test so it passes again.

Fixes: #SYSLOGMT-699

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>
Signed-off-by: László Várady <laszlo.varady@balabit.com>