afsocket: switch to global persist-name duplication check
The persist_hash object was destroyed after the last invocation of the
afsocket_dd_deinit function, but there was no way to detect when the
reference count of a GHashTable reached 0.
For this reason, we could not set the persist_hash pointer to NULL.
After a syslog-ng reload, the dangling persist_hash pointer was dereferenced in
afsocket_dd_init_instance. This could cause segmentation faults.

Switching to the previously introduced "global persist-name duplication check",
we can prevent this dangling pointer problem.

Fixes: #SYSLOGMT-703

Signed-off-by: László Várady <laszlo.varady@balabit.com>