mainloop: Wait for all children processes to finish before shutting down.
Before this patch, the syslog-ng main loop only waited for threads
to finish and it was up to modules to wait for child processes to exit.
However waiting for child processes was not done by for example the indexer
and afprog modules. For the indexer this had two consequences: any message
on the pipe between syslog-ng and the indexer was lost (proven in test).
Also it could happen in theory that an old indexer was still writing files
while a new indexer stating up (in theory, not proven).

This patch introduces a counter that is waited on at shutdown to reach 0.
The counter counts the child processes registered via the children.h child
manager. The counter is increased at registration and decreased if a
process is deregistered or a process SIGCHLD is caught. This means that
the indexer and any other module using the child manager will automatically
wait for its child processes to exit at shutdown. This solves the lost
messages issue and the theoretical race between shutdown and startup of
processes.

Also remove dead code about quit_callback_list.

References: jira SSBM-397