json: integrate format-json with PE template function interface
The TFJsonState data structure has been removed in order to integrate with
PE template functions.

Also, LogTemplateInvokeArgs does not exist in the PE template library,
so it has been replaced with simple parameters.

Note: LogTemplateInvokeArgs should be ported to PE, but first, we need to
redesign the scratch buffer logic in it.

Resolves #SYSLOGDEV-2380

Signed-off-by: László Várady <laszlo.varady@balabit.com>