Logsource: avoid race condition in case of suspend
The main problem is how we handle the queue full event
if flow-control is set, and has multiple destination fed by a source.

It can happen, that the window set to 0 from the destination's thread
This can cause, that the window is set to 0 between calling the log_source_free_to_send
and log_source_queue. This is racy.
This can cause assertion fail when checking that the window size become negative or not.

Anyway, reset the window size to 0 is not a good solution,
because it can be possible, that the window size remain 0 forever

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>