cfg-lexer: add wildcard @include pragma support for Windows
This is the Windows implementation of the cfg_lexer_include_file_glob_at()
function, which is heavily based on the original Unix implementation (with
all of its weird logics).

The function tries to match the given @include pattern. If it fails and
the pattern does not contain a valid wildcard (?, *, []), then the pattern
itself will be added to the include list.

Otherwise all matching file names will be resolved and added to the
include list.

Examples:
- @include "conf.d/?.conf"
- @include "conf.d/*.conf"
- @include "C:\\dir"
- @include "C:/dir"
- @include "C:\\dir/mixed_separator/*.conf"

Fixes #SYSLOGMT-809

Signed-off-by: László Várady <laszlo.varady@balabit.com>