plugin: implement plugin discovery via a separate process
The fact that we are opening modules in the same process as syslog-ng seems
to have a number of ill effects:
  * the thread local storage (TLS) area growing too large, causing
    performance degradation
  * openssl global variables getting reset because we are unloading
    libssl while libcrypt stays in (github #1316)

This patch implements an early version of plugin discovery via a fork()-ed
process. Code is ugly, but I have patches locally that refactor the
plugin stuff and I don't want to introduce major conflicts at this point,
and due to #1316 this seems to be more urgent that the refactor.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>