lib: fix log_center_init_pipe_line() in case of empty rewrite/parser block
Using an empty rewrite or parser block, syslog-ng crashed with a
NULL-pointer dereference.

The log_center_instantiate_process_pipe_line() function was written with a
assumption that the head of the LogProcessRule instance is not NULL, so
this condition must be verified before the instantiation of the process
pipeline.

If the user leaves a rewrite/parser block empty, the head will be, of
course, empty.

This problem has been handled more elegantly in OSE (with the help of
LogExprNodes).

Fixes #SYSLOGDEV-2737

Signed-off-by: László Várady <laszlo.varady@balabit.com>