cfg-lexer: make GlobalConfig dependency explicit
Earlier, the lexer was referring to the actual configuration instance
via the "configuration" global variable. This patch removes the
implicit dependency, and makes that explicit via the constructor.

The coupling between CfgLexer and GlobalConfig is unfortunate and could
probably be improved, however it is still better to see what we
are looking at, rather than hiding it via a global.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>